# condolo/account/src/main/java/com/pichincha/account/domain/external/Customer.java
package com.pichincha.account.domain.external;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.*;
import lombok.experimental.SuperBuilder;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@SuperBuilder
@JsonInclude(JsonInclude.Include.NON_NULL)
public class Customer extends Person {
    private String customerId;
    private String password;
    private boolean status;
}